1.19.4

+++ common/src/main/java/xyz/wagyourtail/minimap/waypoint/Waypoint.java
@@ -278,7 +278,7 @@ public class Waypoint {
     public BlockPos posForCoordScale(double coordScale) {
         return posForCoordScale.computeIfAbsent(coordScale, k -> {
             double scale = this.coordScale / coordScale;
-            return new BlockPos(posX * scale, posY, posZ * scale);
+            return new BlockPos((int) (posX * scale), posY, (int) (posZ * scale));
         });
     }
